# add bigTrend job

+++ soros-data-adaptor/src/main/kotlin/com/soros/data/adaptor/extension/Extensions.kt
@@ -146,7 +146,7 @@ fun StockTrendWaveBo.toBigTrendEntity(dataType: String): BigTrendEntity {
             endDate = endInflectionPoint!!.date,
             waveDirection = waveDirectionEnum.name,
             dataType = dataType,
-            lastDays = lastDays.toLong(),
+            lastDays = lastDays.toBigInteger(),
             range = range
     )
 }
